moving scoring to analyzers/prompt_injection/reward
