+++ llm_defender/core/validators/analyzers/prompt_injection/reward/__init__.py
@@ -1,0 +1,1 @@
+from .scoring import PromptInjectionScoring
