--- llm_defender/core/validators/scoring/__init__.py
@@ -1,1 +1,0 @@
-from . import process
